削除

+++ app/controllers/products_controller.rb
@@ -35,9 +35,10 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
-    product.destroy
-    redirect_to user_menu_index_path
-    
+    if product
+      product.destroy
+    end
+    redirect_to user_menu_index_path 
   end
 
   def user_index
